# [TEST][GPU] rope optimization