--- a/src/plugins/intel_gpu/src/kernel_selector/kernels/rope/rope_kernel_base.cpp
+++ b/src/plugins/intel_gpu/src/kernel_selector/kernels/rope/rope_kernel_base.cpp
@@ -13,6 +13,7 @@ bool RoPEKernelBase::Validate(const Params& p) const {
 JitConstants RoPEKernelBase::GetJitConstants(const rope_params& params, RoPEKernelBase::DispatchData) const {
     JitConstants jit = MakeBaseParamsJitConstants(params);
 
+    jit.AddConstant(MakeJitConstant("VEC_SIZE", params.vec_size));
     jit.AddConstant(MakeJitConstant("HEAD_SIZE", params.head_size));
     jit.AddConstant(MakeJitConstant("ROTARY_NDIMS", params.rotary_ndims));
     jit.AddConstant(MakeJitConstant("HALF_ROTARY_NDIMS", params.rotary_ndims / 2));
@@ -70,23 +71,25 @@ RoPEKernelBase::DispatchData RoPEKernelBase::SetDefault(const rope_params& param
     if (params.is_qwen) {
         dispatchData.gws = {input.Batch().v,
                             input.Feature().v,
-                            params.head_cnt * std::max(params.rotary_ndims / 2ul, params.head_size - params.rotary_ndims)};
+                            params.head_cnt *
+                                std::max(params.rotary_ndims / 2ul, params.head_size - params.rotary_ndims) /
+                                params.vec_size};
     } else if (params.is_chatglm) {
         if (params.support_2d_rope) {
             // input  [batch_size, seq_length]
             // output [batch_size, head_count, seq_length, half_rotary_ndims]
             dispatchData.gws = {input.Batch().v * params.head_cnt,
                                 input.Feature().v,
-                                params.rotary_ndims / 2ul};
+                                params.rotary_ndims / 2ul / params.vec_size};
         } else {
             dispatchData.gws = {input.Batch().v,
                                 input.Feature().v,
-                                params.head_cnt * (params.rotary_ndims / 2ul)};
+                                params.head_cnt * (params.rotary_ndims / 2ul / params.vec_size)};
         }
     } else {
         dispatchData.gws = {output.Batch().v,
                             output.Feature().v,
-                            output.Y().v * params.rotary_ndims / 2ul};
+                            output.Y().v * params.rotary_ndims / 2ul / params.vec_size};
     }
 
     dispatchData.lws = GetOptimalLocalWorkGroupSizes(dispatchData.gws, params.engineInfo, input.GetLayout(), output.GetLayout(), dims_by_gws);