--- a/src/plugins/intel_gpu/src/kernel_selector/kernels/rope/rope_kernel_base.h
+++ b/src/plugins/intel_gpu/src/kernel_selector/kernels/rope/rope_kernel_base.h
@@ -21,6 +21,7 @@ struct rope_params : public base_params {
     size_t axis = 0;
     size_t num_of_inputs = 0;
     size_t gather_rank = 0;
+    size_t vec_size = 1;
 
     bool is_qwen = false;
     bool is_chatglm = false;