… src/tests/functional/shared_test_classes/include/shared_test_classes/subgraph/rotary_pos_emb.hpp
… src/tests/functional/shared_test_classes/src/subgraph/rotary_pos_emb.cpp